fix: carry explicit focus intent through notification dismissal
